+++ src/pages/Blogs/Blogs.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { notify } from "../../utils/notify";
+import { notify , confirmation } from "../../utils/notify";
 import { makeStyles } from "@material-ui/core/styles";
 import Card from "@material-ui/core/Card";
 import CardActionArea from "@material-ui/core/CardActionArea";
@@ -98,17 +98,30 @@ const Blogs = () => {
     window.scrollTo(0, 0);
   };
   const delBlog = (index) => {
-    axios
-      .delete(process.env.REACT_APP_PROD_URL + "api/deleteBlog/" + index)
-      .then((response) => {
-        if (response.data) {
-          notify("success","Successfully Deleted")
-          dispatch(deleteBlog(index));
-        }
-      })
-      .catch((err) => {
-        notify("error","Error in Deleting")
-      });
+    confirmation({
+      title: "Are you sure ? ",
+      message: "Please Confirm",
+      buttons: [
+        {
+          label: "Yes",
+          onClick: () => axios
+          .delete(process.env.REACT_APP_PROD_URL + "api/deleteBlog/" + index)
+          .then((response) => {
+            if (response.data) {
+              notify("success","Successfully Deleted")
+              dispatch(deleteBlog(index));
+            }
+          })
+          .catch((err) => {
+            notify("error","Error in Deleting")
+          })
+        },
+        {
+          label: "No, Cancel",
+        },
+      ],
+    });
+    
   };
 
   return (
