+++ src/pages/Services/Services.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { notify } from "../../utils/notify";
+import { notify, confirmation } from "../../utils/notify";
 import "./Services.css";
 import { DeleteOutline } from "@material-ui/icons";
 import EditIcon from "@material-ui/icons/Edit";
@@ -52,11 +52,14 @@ const Services = () => {
 
     if (editIndex) {
       axios
-        .post(process.env.REACT_APP_PROD_URL + "api/updateService/" + editIndex, {
-          title,
-          description,
-          icon,
-        })
+        .post(
+          process.env.REACT_APP_PROD_URL + "api/updateService/" + editIndex,
+          {
+            title,
+            description,
+            icon,
+          }
+        )
         .then((response) => {
           notify("success", "Successfully Updated");
           dispatch(updateService(response.data));
@@ -88,15 +91,30 @@ const Services = () => {
 
   const handleDelete = (e, index) => {
     e.preventDefault();
-    axios
-      .delete(process.env.REACT_APP_PROD_URL + "api/deleteService/" + index)
-      .then((response) => {
-        notify("success", "Successfully Deleted");
-        dispatch(deleteService(index));
-      })
-      .catch((err) => {
-        notify("error", "Error in Deleting");
-      });
+    confirmation({
+      title: "Are you sure ? ",
+      message: "Please Confirm",
+      buttons: [
+        {
+          label: "Yes",
+          onClick: () =>
+            axios
+              .delete(
+                process.env.REACT_APP_PROD_URL + "api/deleteService/" + index
+              )
+              .then((response) => {
+                notify("success", "Successfully Deleted");
+                dispatch(deleteService(index));
+              })
+              .catch((err) => {
+                notify("error", "Error in Deleting");
+              }),
+        },
+        {
+          label: "No, Cancel",
+        },
+      ],
+    });
   };
 
   return (
